Show the same monitoring links in the ClusterMetrics component
Previously, the Cluster Metrics block showed different links based on the user's
role: Gardener operator (`isAdmin`) or regular shoot owner.

Shoot owners saw a link to the "users" Grafana instance and optionally to an
alert manager instance.

Gardener operators saw a link to the "operators" Grafana instance, to a
Prometheus and optionally to an alert manager instance.

With the Gardener PR 7007, both shoot owners and Gardener operators shall have
equal access to the observability components. The two different Grafana
instances are replaced by a single Grafana instance (that keeps the old name,
"users" Grafana) that is accessed by both roles.

With this commit, both shoot owners and Gardener operators will see the same
monitoring links: a link to a Grafana (the "users" Grafana), a link to a
Prometheus and optionally a link to an alert manager instance.

The password that used to be shown to shoot owners will also be shown to
Gardener operators. With Gardener PR 7007, the shoot owner password will grant
access to the Prometheus instance as well.

Previously the seedInfo feature was used to get the password for Gardener
operators from the shoot's control plane in the seed. It is no longer needed and
will be removed.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>