Remove no longer needed code from assignMonitoringSecret
Previously, assignMonitoringSecret used to handle both cases: get the monitoring
credentials for shoot owners from the virtual-garden cluster and for Gardener
operators from the seed cluster.

Now, the "user" monitoring secret should be used for both roles so the logic to
access the monitoring-ingress-credentials in the seed can be removed.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>