Remove the seed info feature from the frontend
The seedInfo was used to get the monitoring password for the Gardener operators.
With this PR, Gardener operators will use the same password as shoot owners and
hence the seedInfo is no longer needed. There are no other usages of the
seedInfo in the frontend, so the seedInfo feature can be removed entirely.

The probably obsolete TODO containing the word "seedInfo":
> on shootsubscription done do not fetch seedInfo
is also removed.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>