Let shoots.info() always return the user monitoring secret
Previously, shoot.info only returned the monitoring secret if the user was not
the operator. For the operator, the monitoring secret was returned by the
previously removed seed-info API.

The mock setup of the acceptance test had to be adjusted because the
`authorization.isAdmin(user)` call is no longer executed and hence there is no
self subject access review in the code flow.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>