Adjust the shoot ingress domain calculation
With Gardener 1.64, a bug was fixed that affected technical IDs and the
ingress calculations.

In older shoot clusters with a single - in the technical ID, the ingress
domain contained the prefix 'shoot-' (this was a bug), because Gardener
tried to remove "shoot--" and that was not found so the "shoot-" prefix
stayed in the ingress domain.

Now that the bug is fixed, the 'shoot-' prefix will be removed and the
logic in the Gardener Dashboard has to be adjusted as well.

See https://github.com/gardener/gardener/pull/7454

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Holger Koser <holger.koser@sap.com>