Shoot owners and Gardener operators see the same monitoring links (#1408)
* Show the same monitoring links in the ClusterMetrics component

Previously, the Cluster Metrics block showed different links based on the user's
role: Gardener operator (`isAdmin`) or regular shoot owner.

Shoot owners saw a link to the "users" Grafana instance and optionally to an
alert manager instance.

Gardener operators saw a link to the "operators" Grafana instance, to a
Prometheus and optionally to an alert manager instance.

With the Gardener PR 7007, both shoot owners and Gardener operators shall have
equal access to the observability components. The two different Grafana
instances are replaced by a single Grafana instance (that keeps the old name,
"users" Grafana) that is accessed by both roles.

With this commit, both shoot owners and Gardener operators will see the same
monitoring links: a link to a Grafana (the "users" Grafana), a link to a
Prometheus and optionally a link to an alert manager instance.

The password that used to be shown to shoot owners will also be shown to
Gardener operators. With Gardener PR 7007, the shoot owner password will grant
access to the Prometheus instance as well.

Previously the seedInfo feature was used to get the password for Gardener
operators from the shoot's control plane in the seed. It is no longer needed and
will be removed.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

* Adjust the Grafana URL variables

- remove the no longer used grafanaUrlOperators variable
- rename the grafanaUrlUsers variable to grafanaUrl

Note that the "gu" prefix of the users Grafana is kept for the new single
Grafana instance (also in Gardener PR 7007), so that the Grafana URL that
is shown by the Gardener Dashboard will be valid during the transition
period: when the new version of the Gardener Dashboard is already deployed
but the shoot cluster has not reconciled yet.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

* Remove the seed info feature from the frontend

The seedInfo was used to get the monitoring password for the Gardener operators.
With this PR, Gardener operators will use the same password as shoot owners and
hence the seedInfo is no longer needed. There are no other usages of the
seedInfo in the frontend, so the seedInfo feature can be removed entirely.

The probably obsolete TODO containing the word "seedInfo":
> on shootsubscription done do not fetch seedInfo
is also removed.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

* Remove the seed info API from the backend

The corresponding acceptance tests are also removed.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

* Let shoots.info() always return the user monitoring secret

Previously, shoot.info only returned the monitoring secret if the user was not
the operator. For the operator, the monitoring secret was returned by the
previously removed seed-info API.

The mock setup of the acceptance test had to be adjusted because the
`authorization.isAdmin(user)` call is no longer executed and hence there is no
self subject access review in the code flow.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

* Remove no longer needed code from assignMonitoringSecret

Previously, assignMonitoringSecret used to handle both cases: get the monitoring
credentials for shoot owners from the virtual-garden cluster and for Gardener
operators from the seed cluster.

Now, the "user" monitoring secret should be used for both roles so the logic to
access the monitoring-ingress-credentials in the seed can be removed.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

---------

Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>