Replace references to Grafana and Loki with Plutono and Vali (#1494)
In https://github.com/gardener/gardener/pull/7318 we replaced Grafana and
Loki with Plutono/Vali. These changes should be reflected in the dashboard.
The ingress URL is unchanged for now, because there is a separate migration
planned for that. This PR is about textual consistency and does not change
any functionality or behavior. It therefore does not require a simultaneous
release with the g/g changes.

If you want to learn more about the motivation for this change, please
refer to the PR above and associated issues.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>